--- app/blog/(posts)/rust-backend-memo/page.tsx
@@ -1,2 +1,0 @@
-export { metadata, post } from "./[id]";
-export { default } from "./[id]";
