--- app/blog/(posts)/study-note-growth-mindset/page.tsx
@@ -1,2 +1,0 @@
-export { metadata, post } from "./[id]";
-export { default } from "./[id]";
